+++ CHANGELOG.rst
@@ -7,6 +7,7 @@ Changelog
 * Add a new genotyping platform, ``LongRead``, to the :command:`import-variants` command.
 * Add new command :command:`run-long-read-pipeline`.
 * Remove the ``Code`` column from the ``cnv-table.csv`` file. From now on, CNV codes will be generated on the fly.
+* Add new method :meth:`api.core.load_cpic_table`.
 
 0.12.0 (2022-01-29)
 -------------------
@@ -28,7 +29,7 @@ Changelog
 * Fix minor bug in :command:`compute-copy-number` command.
 * Update :command:`plot-cn-af` command to check input files more rigorously.
 * Improve CNV caller for CYP2A6, CYP2D6, and SLC22A2.
-* Add new method :meth:`sdk.utils.add_cn_samples` method.
+* Add new method :meth:`sdk.utils.add_cn_samples`.
 * Update :command:`compare-genotypes` command to output CNV comparisonw results as well.
 * Update :command:`estimate-phase-beagle` command. From now on, the 'chr' prefix in contig names (e.g. 'chr1' vs. '1') will be automatically added or removed as necessary to match the reference VCF’s contig names.
 * Add index files for 1KGP reference haplotype panels.